cloud_functions/token_refresh/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-pubsub==2.*
stravalib==1.3.0
google-cloud-secret-manager==2.*
pydantic==1.10.10